need update-with-limit reintroduced, but fetch_tasks is now operational
